requirements.txt
streamlit-1.35.0-py2.py3-none-any.whl
Authlib